requirements.txt
langchain_experimental~=0.3.4
langchain~=0.3.26
beautifulsoup4~=4.13.4
requests~=2.32.4
pandas~=2.3.0
regex~=2024.11.6
underthesea~=6.8.4
PyMySQL~=1.1.1
qdrant-client~=1.14.3
python-dotenv~=1.1.1
tqdm~=4.67.1
langchain-text-splitters~=0.3.8
torch~=2.6.0+cu124
langchain-google-genai~=2.1.6
protobuf~=6.31.1
google~=3.0.0
openai~=1.93.0
langchain-core~=0.3.68
rank-bm25~=0.2.2
transformers~=4.53.1
flask~=3.1.1
flask-cors~=6.0.1
werkzeug~=3.1.3
selenium~=4.34.2
webdriver-manager~=4.0.2
langchain-community~=0.3.27
pydantic~=2.11.7
opensearch-py~=3.0.0
cohere~=5.16.1